+++ packages/http-core/src/interfaces/HttpResponse.ts
@@ -9,4 +9,5 @@ export interface HttpResponse<M extends ResponseData> {
   requestId: string;
   correlationId: string;
   timestamp: number;
+  reasonCode?: string;
 }