+++ packages/fastify-core/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2025-09-23
+
+### Added
+- Add `reasonCode` to response
+
+## [2.3.4] - 2025-09-23
+
 ## [2.3.3] - 2025-09-16
 
 ### Changed