+++ packages/http-client/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.0] - 2025-09-23
+
+### Added
+- Return `reasonCode` in response
+
 ## [3.0.3] - 2025-09-16
 
 ### Changed