+++ packages/http-core/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to the `@fistware/http-core` package will be documented in t
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2025-09-23
+
+### Added
+- Added optional `reasonCode` field to `HttpResponse` interface
+
 ## [3.0.1] - 2025-08-23
 
 ### Added